Allowed the json data to be read encoded as UTF-8
To resolve the encoding issue with need to reload our initial data.
The json data currently on RavenDB has the encoding problems in it.
This was due to the fact that the System.Net.WebClient class doesn't
read UTF-8 by default. This commit fixes that issue.

Please, reload the RavenDB instance we are using with the new
ShindyDataLoader.